Minor improvements in TreePicker
Basically, added a Fragment as a root for the first map iteration, this fixes an error
shown in the developer console caused by having a list without a proper key element.

Also, added a Tooltip for the ListItems. If the screen is too small, the text might be
cropped, and the name wouldn't be clear, so hovering the text will also display the name of the bucket