Start migrating to plugin-api services
Only those services that are already using the new reference instead of the old ones.
As some plugins are still in progress we have to use the parser to winston to avoid errors, but at least
this new implementation won't cause errors when the migration is needed.

This change allowed us to remove the dependency in winston