Fix config.d.ts for allowedBuckets
The @backstage/backend-plugin-api package had a bunch of breaking changes, and this was
causing issues with the type `Record<string, string[]>` we were using for the allowedBuckets.

Due to that, we needed to change the configuration type. It just works the same ways as before,
but now it's just an array which defines a platform and a list of buckets. I updated the docs,
the config.d.ts as well as added a breaking change line in the changeset.

If you wonder, that seems to be the expected behavior now. The proxy had a similar format
and it was updated as well in this release.