Fix package metadata
During the publish process, the pipeline failed because of a recent
change in the way the packages are published. In concrete, with the command
"yarn backstage-cli package prepack".

This also happened to changes done upstream like the vault and grafana plugins.
It's something recent and seems like they are being fixed on the fly when
they occur.