Fix local dependencies for plugins
This has been done by using the [following script](https://github.com/backstage/backstage/blob/master/scripts/verify-local-dependencies.js) available
in the main Backstage repo.

This is also the way the community-plugin workspaces work, so it's good to follow their setup too.

It may also help with the issue we had recently in the publish step