Implement testing environment
After removing the old setup, we can now add a new setup like it's done
in the community plugins. This is based in creation a kind of plain
full application which only has the needed pieces for us.

In our case, we have the catalog (mostly for authentication), the s3-viewer
and the settings. The permissions are also installed in case something
has to be tested. By default all requests are accepted.