Fix releases for plugins
The plugins were not able to be discovered anymore due to the
workspace usage we implemented in the last PRs. Let's try reverting
that change and make sure the packages can be found again.

This could potentially lead to the releases needing a local `yarn install`
to make sure the pipelines pass. But that's better than nothing.