Don't unconditionally add `download` property to "Download" tags.
This defeats the changes in adfee42c4dd2aae25d11909013beb6e4a6495210.

Instead rely on the presence of the "Content-Disposition: attachment"
header for whether or not the browser will download or open the linked
file.

This allows viewing html files as web pages in the browser rather than
having the browser download them attachments, which is the experience
that users prefer for html by default.